Only track unordered receive state when required for unordered reads
Ordered stream allow simpler, faster logic to determine whether a
stream has been fully read because we are guaranteed that
discontinuities are represented by still-buffered data. Only when
unordered reads are in use do we need to handle the case for a gap
between stream frames which we no longer store in any form.